requirements.txt
azure-functions
google-auth
google-auth-httplib2
google-api-python-client
python-dotenv
fitz
PyMuPDF
openai